Exporter changes for handling RAS events on K8S
Exporter now exposes an interface that the AICC can use
to update the compute node's health status upon receiving
RAS events. Internally, based on the updated health state,
the Exporter will automatically mark the GPUs as either
available or unavailable in it's health state which the
device plugin will use to mark the GPUs state.